+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Fixed detecting computed properties
 - Fixed typo in `isConvenienceInitialiser` property
 - Fixed creating cache folder when cache is disabled
+- Fixed parsing multiple enum cases annotations
 
 ### Internal changes